Project restructuring.  Java code is now for example purposes.  Made bindings folder for all Makefile and code related to langage bindings
