+++ bindings/java/java-p2pc.h
@@ -1,0 +1,14 @@
+#include <jni.h>
+
+#ifndef _Included_p2pc
+#define _Included_p2pc
+#ifdef __cplusplus
+extern "C" {
+#endif
+
+JNIEXPORT void JNICALL Java_p2pc_startClient (JNIEnv *, jclass, jstring);
+
+#ifdef __cplusplus
+}
+#endif
+#endif
