+++ bindings/main.go
@@ -2,15 +2,11 @@ package main
 
 import "C"
 import (
-	identify "github.com/libp2p/go-libp2p/p2p/protocol/identify"
-
+	p2pd "github.com/libp2p/go-libp2p-daemon"
 	p2pc "github.com/libp2p/go-libp2p-daemon/p2pclient"
 )
 
 func main() {
-	identify.ClientVersion = "p2pc/0.1"
-	config := p2pc.Initialize()
-	p2pc.Start(config)
 }
 
 //export startClient
@@ -18,4 +14,16 @@ func startClient(args *C.char) {
 	argsGoString := C.GoString(args)
 	config := p2pc.ProcessArgs(&argsGoString)
 	p2pc.Start(config)
+}
+
+//export startDaemon
+func startDaemon(args *C.char) {
+	argsGoString := C.GoString(args)
+	config := p2pd.ProcessArgs(&argsGoString)
+	p2pd.Start(config)
+}
+
+//export stopDaemon
+func stopDaemon() {
+	p2pd.Stop()
 }
